Fixed A* + got heroes to face direction they're running.

+++ trunk/src/Structures/Graph/GraphEdge.cpp
@@ -2,7 +2,7 @@
 
 using namespace Structures;
 
-GraphEdge::GraphEdge(const GraphNodePtr &from, const GraphNodePtr &to, unsigned int cost)
+GraphEdge::GraphEdge(const GraphNodePtr &from, const GraphNodePtr &to, float cost)
   : from(from), to(to), cost(cost)
 {
 }
